Make c++ ffi work

+++ Day20v2/mix.h
@@ -4,7 +4,7 @@
 extern "C" {
 #endif
 
-extern int *mix(int *coords, int length, int rounds);
+extern long long int *mix(long long int *coords, int length, int rounds);
 
 #ifdef __cplusplus
 }
